chore(flexnative-demo): update package-lock.json and refine .gitignore
This commit updates the `package-lock.json` of the `flexnative-demo` project with new dependencies, including @flexnative/authentication and libraries like axios, formik, jwt-decode, and yup. It also amends the `.gitignore` file to exclude build and configuration files specific to certain directories, such as `npm-packages`, and `flexnative-demo`. These changes ensure that temporary files and lock files that should not be tracked are ignored, while aligning dependencies with the current development requirements.